add webpack for bundling
