+++ common/index.js
@@ -6,20 +6,8 @@ function getConcealedText(config, originalText) {
 
 	var concealedText = originalText.slice();
 
-	var concealedText = originalText.replace(/[\u{0080}-\u{FFFF}]/gu,"");
+	const transformation = transformations.createTransformation(config);
+	const concealed = transformation.getEncoded(config, originalText);
 
 	return concealedText;
 }
-
-
-//console.log(transformations);
-const asd = transformations.Encrypt.getEncoded({password: "asd"},"aaaaaaaaaaaaaaaaaaaaaaaaaaa");
-//console.log(asd);
-//console.log(transformations.Encrypt.getDecoded({password: "asd"}, asd));
-const config = {password:"a", removeNonASCII:true, compressPlainText:true, encrypt:true};
-const t = transformations.createTransformation(config);
-const message = "Én vagyok a Betmenn !!%56";
-console.log(message);
-const concealed = t.getEncoded(config, message);
-console.log(concealed);
-console.log(t.getDecoded(config, concealed));
